+++ src/Krystal/Validate/Pattern/Collection.php
@@ -1,0 +1,30 @@
+<?php
+
+/**
+ * This file is part of the Krystal Framework
+ * 
+ * Copyright (c) No Global State Lab
+ * 
+ * For the full copyright and license information, please view
+ * the license file that was distributed with this source code.
+ */
+
+namespace Krystal\Validate\Pattern;
+
+final class Collection extends AbstractPattern
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function getDefinition()
+    {
+        return $this->getWithDefaults(array(
+            'required' => true,
+            'rules' => array(
+                'NotEmpty' => array(
+                    'message' => 'Pick at least one element in collection'
+                )
+            )
+        ));
+    }
+}